refactor(lld): give the emulated screen buttons a view model
