+++ apps/ledger-live-desktop/src/mvvm/components/DeviceScreen/components/DeviceScreenButtons/index.tsx
@@ -1,0 +1,14 @@
+import React from "react";
+import { DeviceScreenButtonsView } from "./DeviceScreenButtonsView";
+import {
+  useDeviceScreenButtonsViewModel,
+  type DeviceScreenButtonPress,
+} from "./useDeviceScreenButtonsViewModel";
+
+export interface DeviceScreenButtonsProps {
+  readonly onPress: DeviceScreenButtonPress;
+}
+
+export function DeviceScreenButtons({ onPress }: DeviceScreenButtonsProps) {
+  return <DeviceScreenButtonsView {...useDeviceScreenButtonsViewModel(onPress)} />;
+}
